Add wrapper function for tree to zshrc
This colorizes and auto-paginates (only when needed) the output of
`tree`.

I mentioned this to @nurugger07 at rubyjax, and offered to submit a pull
request for your consideration. I have not actually tested this exact
configuration on a live machine (as I don't want to install dotmatrix),
but the configuration is pulled straight from my zshrc, so it *should*
work.