Add battery command
This is general-purpose battery percentage indicator for laptops. There
are quite a few formatters included, and several options.

It is known to work with Ubuntu and Mac OS, and even works on Ubuntu
running through Vagrant/VirtualBox (my use case).

The --tmux option is useful for tmux status bar output. For example, in
my current tmux config, I have the following:

    set -g status-right ' #(hr battery --format="[%%percent]" --charging --color --tmux)'