Set the PGDATABASE env var when hcd is used
When opening `psql` with the PGDATABASE env var set
the connected database will be the name in the env
var.

This feature relies on the convention of having a database
with the same name as the project but with `_development` as
a suffix.

When hcd is used, the PGDATABASE env var is set and
we acquire the convienience of only ever needing to type
`psql`.