Remove NeoFormat - Use ALE for formatting
Since ALE is capable of autoformatting code and linting there is no need
for NeoFormat.

ALE's autoformatting also allows using lambda functions and allows to
easily disable autoformatting for a buffer which comes in handy when
running `:Btabedit`.

This also removes the forced prettier configuration that was hardcoded
into vimrc. Instead place a .prettierrc in your $HOME folder which you
can override with a .prettierrc file in your project dir.