Add :Driver command to re-source user specific rc
Often we have idiosyncrasies in how we work. We encode these into our
muscles and having to deal with pairing situations where you are out of
sync with your pair is painful sometimes. By adding a convenience
function around sourcing your own set of configurations, we might
alleviate some of this frustration.